+++ src/Precompile.jl
@@ -50,7 +50,7 @@ function __init__()
         :a => Symbol => [:block, :mode]
         text"[\"a\"]" => Text => [:mode]
         (true, 1, 1.2, 1 // 2) => Tuple => [:mode]
-        TypstFunction(context, typst"") => TypstFunction => [:depth, :indent, :mode, :tab_size]
+        TypstFunction(Symbol("")) => TypstFunction => [:depth, :indent, :mode, :tab_size]
         typst"[\"a\"]" => TypstString => Symbol[]
         TypstText([1, 2, 3, 4]) => TypstText => Symbol[]
         Typst(1) => Typst => Symbol[]