+++ src/contexts/ContextErrors.jl
@@ -46,14 +46,14 @@ show(io::IO, ::MIME"text/plain", ce::ContextError) = print(
     io, ContextError, "(", ce.expected, ", ", ce.received, ", :", ce.key, ")"
 )
 
-function _unwrap(data_type::DataType, key::Symbol, value)
+function _unwrap(data_type::Type, key::Symbol, value)
     value isa data_type ? value : throw(ContextError(data_type, typeof(value), key))
 end
 
 unwrap(collection, key::Symbol, default) = _unwrap(
     typeof(default), key, get(collection, key, default)
 )
-function unwrap(collection, type::DataType, key)
+function unwrap(collection, type::Type, key)
     value = collection[key]
     _unwrap(type, key, value)
 end