+++ src/strings/TypstFunctions.jl
@@ -1,0 +1,8 @@
+
+module TypstFunctions
+
+import Base: ==, show
+
+using ..Strings: AbstractTypst, TypstString
+
+end # TypstFunctions