+++ src/strings/TypstTexts.jl
@@ -1,0 +1,8 @@
+
+module TypstTexts
+
+import Base: ==
+
+using ..Strings: AbstractTypst
+
+end # TypstTexts