+++ src/strings/Typsts.jl
@@ -1,0 +1,8 @@
+
+module Typsts
+
+import Base: ==
+
+using ..Strings: AbstractTypst
+
+end # Typsts